+++ dipylonfile.py
@@ -29,19 +29,79 @@ import cgi
 from gaps import Gaps
 import xml.etree.ElementTree as ElementTree
 
-LISTOFNOTESCATEGORIES = ["grammar",]
+LISTOFNOTESCATEGORIES = ["word",
+                         "words",
+                         "extract"]
+
+class ListOfNotes(list):
+    """
+        List of Notes objects
+    """
+
+    #///////////////////////////////////////////////////////////////////////////
+    def __init__(self):
+        list.__init__(self)
+
+    #///////////////////////////////////////////////////////////////////////////
+    def get_color_for_one_index(self):
+        """
+                <self> is a list of notes relative to one character.
+        """
+        return 0x990000
+
+    #///////////////////////////////////////////////////////////////////////////
+    def get_text_segments(self):
+        """
+                return a list of (colored) (text) segments
+        """
+        characters = {} # index : [list of Note objects]
+        for note in notes:
+
+            for index in note.gaps.get_all_indexes_one_by_one():
+                if not index in characters:
+                    characters[index] = ListOfNotes()
+                else:
+                    characters[index].append( note )
+
+        colored_characters = {} # index : color
+        for index in characters:
+            colored_characters[index] = characters[index].get_color_for_one_index()
+        indexes_in_colored_characters = sorted(colored_characters.keys())
+
+        # we bring together the characters with the same color :
+        res = {}
+        current_gap = []
+        for index in indexes_in_colored_characters:
+            if current_gap != []:
+                current_gap.append(index)
+            elif index == min(current_gap)-1:
+                pass
+            elif index == max(current_gap)+1:
+                pass
+            else:
+                current_gap = []
+                res[ current_gap ]
+            
+        
+        
+        return res
+        
 
 ################################################################################
 class Note(object):
-    def __init__(self, gaps, corresponding_text, note):
+    def __init__(self, gaps, corresponding_text, note, category, aspect):
         self.gaps = gaps
         self.corresponding_text = corresponding_text
         self.note = note
+        self.category = category
+        self.aspect = aspect
 
     def __repr__(self):
-        return "(Note) gaps={0}; corresponding_text={1}; note={2}".format(self.gaps,
-                                                                          self.corresponding_text,
-                                                                          self.note)
+        msg = "(Note) gaps={0}; corresponding_text={1}; note={2}; aspect={3}"
+        return msg.format(self.gaps,
+                          self.corresponding_text,
+                          self.note,
+                          self.aspect)                                                                           
 
 ################################################################################
 class DipylonFile(object):
@@ -59,7 +119,7 @@ class DipylonFile(object):
                 to a certain <category>. If <category> is equal to None, all
                 categories may be added to the result.
         """
-        res = []
+        res = ListOfNotes()
         
         if category is not None:
             _categories = (category,)
@@ -76,7 +136,10 @@ class DipylonFile(object):
                     
                     res.append( Note( gaps = gaps,
                                       corresponding_text = note['corresponding_text'],
-                                      note = note['note'] ))
+                                      note = note['note'],
+                                      category = _category,
+                                      aspect = note['aspect'],
+                                      ))
 
         return res
 
@@ -143,7 +206,11 @@ class DipylonFile(object):
                     gaps = Gaps().init_from_xmlrepresentation( note.attrib["gaps"] )
                     corresponding_text = note.attrib["corresponding_text"]
                     category = note.attrib["category"]
+                    aspect = note.attrib["aspect"]
+                    
                     self.datalanguages[ language_name ]["notes"][category][gaps.get_xmlrepresentation()] = {"note" : note.text,
-                                                                                                            "corresponding_text" : corresponding_text}
+                                                                                                            "corresponding_text" : corresponding_text,
+                                                                                                            "aspect" : aspect,
+                                                                                                            }
 
         return self